refactor helper fns to utils file
